Add tests for stack and queue, rename folders following camel case
