+++ Src/DataStructures/Src/Stack.c
@@ -1,4 +1,4 @@
-#include "../include/StackADT.h"
+#include "../Include/StackADT.h"
 #include <stdlib.h>
 
 Stack* stack_init(void (*free_function)(void *)) {
