+++ Tests/Include/QueueTests.h
@@ -1,0 +1,5 @@
+void queue_simple_test();
+
+void queue_complex_test();
+
+void queue_overall_test();
