--- tests/UnityUnitTests/src/unity.c:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\Admin\Downloads\Unity-master.zip
