--- tests/include/QueueTests.h
@@ -1,1 +1,0 @@
-void queue_unit_tests();
